Route shutdown + single-instance through Platform trait
Per-OS process-lifecycle code (SIGINT/SIGTERM handlers, single-instance
socket/pipe) moves out of the shared app crate behind new Platform trait
methods: unix defaults in jfn_platform_abi, Windows overrides in jfn_windows.
The callback and instance id are injected so the abstraction crate gains no
new crate dependencies.
