+++ src/jfn_rust/Cargo.toml
@@ -41,14 +41,6 @@ jfn-macos = { path = "../macos" }
 
 [target.'cfg(target_os = "windows")'.dependencies]
 jfn-windows = { path = "../windows" }
-windows-sys = { workspace = true, features = [
-    "Win32_System_Threading",
-    "Win32_System_Pipes",
-    "Win32_System_IO",
-    "Win32_Storage_FileSystem",
-    "Win32_Foundation",
-    "Win32_Security",
-] }
 
 [target.'cfg(target_os = "windows")'.build-dependencies]
 embed-resource = "3"
