+++ src/jfn_rust/src/lib.rs
@@ -1,10 +1,8 @@
 pub mod app;
 mod cli;
+mod instance_id;
 pub mod manager;
 mod platform_install;
-#[cfg(unix)]
-pub mod signal_guard;
-mod single_instance;
 mod window_geometry;
 #[cfg(target_os = "linux")]
 pub mod wl_interpose;
